requirements.txt
amqp==5.3.1
annotated-types==0.7.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clickhouse-connect==0.14.1
idna==3.11
kafka-python==2.3.0
kombu==5.6.2
lz4==4.4.5
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pmxt==2.21.1
prompt_toolkit==3.0.52
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
redis==7.3.0
requests==2.32.5
six==1.17.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zstandard==0.25.0
